first part almost ok

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: halnuma <halnuma@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 16:17:58 by halnuma           #+#    #+#             */
-/*   Updated: 2024/11/06 15:35:10 by halnuma          ###   ########.fr       */
+/*   Updated: 2024/11/07 13:16:18 by halnuma          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
